MathViz/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
manim==0.18.1
groq==0.12.0
bcrypt==4.0.1
python-jose==3.3.0
gunicorn==21.2.0
numpy>=1.19.0
pillow>=7.0.0
pycairo>=1.19.0
pygments>=2.6.1
rich>=6.0.0
